cambios para la adaptacion del hsql

+++ src/main/java/ar/edu/unlam/scaw/modelo/Usuario.java
@@ -13,15 +13,15 @@ public class Usuario {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private int id;
 	private String email;
 	private String password;
 	private String rol;
 	
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getEmail() {
